+++ 06_Algo_QuickAndMergeSort/A_MergeSort_recursion.cpp
@@ -1,10 +1,10 @@
 #include <iostream>
 
-void mergeArr(int *arr, size_t start1, size_t end1, size_t start2, size_t end2) {
-    size_t start = start1;
-    size_t end = end2;
-    size_t index = 0;
-    size_t *res = new size_t[end1 + end2 + 2];
+void mergeArr(int *arr, int start1, int end1, int start2, int end2) {
+    int start = start1;
+    int end = end2;
+    int index = 0;
+    int *res = new int[end1 + end2 + 2];
 
     while (start1 <= end1 && start2 <= end2) {
         if (arr[start1] < arr[start2]) {
@@ -18,19 +18,19 @@ void mergeArr(int *arr, size_t start1, size_t end1, size_t start2, size_t end2) 
     }
 
     if (start1 <= end1) {
-        for (size_t i = start1; i <= end1; ++i) {
+        for (int i = start1; i <= end1; ++i) {
             res[index] = arr[i];
             index++;
         }
     } else {
-        for (size_t i = start2; i <= end2; ++i) {
+        for (int i = start2; i <= end2; ++i) {
             res[index] = arr[i];
             index++;
         }
     }
 
     index = 0;
-    for (size_t i = start; i <= end; ++i) {
+    for (int i = start; i <= end; ++i) {
         arr[i] = res[index];
         index++;
     }
@@ -38,8 +38,8 @@ void mergeArr(int *arr, size_t start1, size_t end1, size_t start2, size_t end2) 
     delete[] res;
 }
 
-void mergeSort(int *arr, size_t first, size_t last) {
-    size_t mid = first + (last - first) / 2;
+void mergeSort(int *arr, int first, int last) {
+    int mid = first + (last - first) / 2;
 
     if (first < last) {
         mergeSort(arr, first, mid);
@@ -49,17 +49,17 @@ void mergeSort(int *arr, size_t first, size_t last) {
 }
 
 int main() {
-    size_t n;
+    int n;
     std::cin >> n;
     int *arr = new int[n];
 
-    for (size_t i = 0; i < n; ++i) {
+    for (int i = 0; i < n; ++i) {
         std::cin >> arr[i];
     }
 
     mergeSort(arr, 0, n - 1);
 
-    for (size_t i = 0; i < n; ++i) {
+    for (int i = 0; i < n; ++i) {
         std::cout << arr[i] << " ";
     }
     std::cout << std::endl;
